Create oct28.448pm
what does int arr[2][2] mean in main()
The declaration int arr[2][2] in the function main() means that an integer array with two rows and two columns has been declared.  The compiler will allocate memory for the array, and the name arr will be a pointer to the first element of the array.

The syntax for declaring a two-dimensional array in C++ is as follows:

<data type> <array name>[<number of rows>][<number of columns>]
In this case, the data type is int, the array name is arr, the number of rows is 2, and the number of columns is 2.

Therefore, the declaration int arr[2][2] in the function main() creates a two-dimensional array with two rows and two columns.  The memory for the array is allocated by the compiler, and the name arr is a pointer to the first element of the array.

This is a typical way to declare a multidimensional array in C++. It allows the programmer to store a collection of data in a single, organized structure. The array can then be accessed and manipulated using a variety of operators and functions.