Create oct28.507pm
what does void processArr(int a[] [2] mean in the void function
View other drafts


















The void function processArr(int a[][2]) means that the function takes an integer array with two rows and two columns as its argument and does not return any value.

The void keyword specifies that the function does not return any value. The int keyword specifies that the function takes an integer array as its argument. The a[][] syntax specifies that the function takes a two-dimensional array as its argument. The number of rows and columns in the array are specified by the two integer values in the brackets.

The function processArr can be used to perform any operation on the array that is passed to it. For example, the function could be used to print the elements of the array to the console, to calculate the sum of the elements of the array, or to sort the elements of the array.

Here is an example of how the function processArr could be used:

C++
void processArr(int a[][2]) {
  for (int i = 0; i < 2; i++) {
    for (int j = 0; j < 2; j++) {
      cout << a[i][j] << " ";
    }
    cout << endl;
  }
}

int main() {
  int arr[2][2];
  arr[0][0] = 0;
  arr[0][1] = 1;
  arr[1][0] = 2;
  arr[1][1] = 3;

  processArr(arr);

  return 0;
}
Use code with caution. Learn more
Output:

0 1
2 3
The function processArr is a very versatile function that can be used to perform a variety of operations on a two-dimensional integer array.