Create OCT28C10603PM
how is the variable student an element parameter in the scores array....in ....#include <iostream>

int main()
{
    int scores[]{ 84, 92, 76, 81, 56 };
    int numStudents = 5;

    int maxScore = 0; // keep track of our largest score
    for (int student = 0; student < numStudents; ++student)
    {
        if (scores[student] > maxScore)
        {
            maxScore = scores[student];
        }
    }

    std::cout << "The best score was " << maxScore << '\n';

    return 0;
}
View other drafts





















1
The variable student is an element parameter in the scores array because it is used to index into the array. When the code scores[student] is executed, the value of the variable student is used to determine which element of the scores array to access.

For example, when the code scores[0] is executed, the value of the variable student is 0, which means that the first element of the scores array is accessed. When the code scores[1] is executed, the value of the variable student is 1, which means that the second element of the scores array is accessed, and so on.

The variable student is called an element parameter because it is used to select which element of the array to access. This is in contrast to a range parameter, which is used to specify a range of elements in the array to access.

In the code you provided, the variable student is used as an element parameter in the for loop. The for loop iterates over the scores array, and the variable student is used to index into the array and access each element.

Here is a breakdown of the code:

C++
int scores[]{ 84, 92, 76, 81, 56 };
int numStudents = 5;

int maxScore = 0; // keep track of our largest score

for (int student = 0; student < numStudents; ++student)
{
        if (scores[student] > maxScore)
        {
            maxScore = scores[student];
        }
    }

    std::cout << "The best score was " << maxScore << '\n';

    return
 
0;
}
Use code with caution. Learn more
The first line declares an array of integers called scores and initializes it with the values 84, 92, 76, 81, and 56.
The second line declares an integer variable called numStudents and initializes it with the value 5.
The third line declares an integer variable called maxScore and initializes it with the value 0.
The fourth line starts a for loop.
The fifth line declares an integer variable called student and initializes it with the value 0.
The sixth line is the condition for the for loop. The loop will iterate as long as the value of student is less than the value of numStudents.
The seventh line increments the value of student by 1.
The eighth line is the body of the for loop. The code inside the body of the for loop will be executed for each element of the scores array.
The ninth line is an if statement. If the value of scores[student] is greater than the value of maxScore, then the value of maxScore is set to the value of scores[student].
The tenth line ends the if statement.
The eleventh line ends the for loop.
The twelfth line prints the value of maxScore to the console.
The thirteenth line returns 0 from the main() function.
The variable student is an element parameter in the scores array because it is used to index into the array. The for loop iterates over the scores array, and the variable student is used to index into the array and access each element.