Create OCT29.356PM
WHAT DOES THE COLON MEAN IN ...for (int number : fibonacci) // iterate over array fibonacci
View other drafts




















The colon in the code snippet for (int number : fibonacci) // iterate over array fibonacci is the range operator. It is used to iterate over the elements of an array or other iterable object.

The range operator takes two operands: a variable and an iterable object. The variable is assigned the value of each element of the iterable object in turn. In the code snippet above, the variable number is assigned the value of each element of the array fibonacci.

The range operator is a very powerful tool for iterating over collections of data. It is used in many different programming languages, including Java, Python, and C++.

Here is an example of how the range operator can be used to iterate over a list of numbers:

List<int> numbers = Arrays.asList(1, 2, 3, 4, 5);

for (int number : numbers) {
    System.out.println(number);
}
Output:

1
2
3
4
5
You can also use the range operator to iterate over the elements of a string:

String str = "Hello, world!";

for (char character : str) {
    System.out.print(character);
}
Output:

Hello, world!
The range operator is a very versatile tool that can be used in many different ways. It is a good tool to learn if you are serious about programming.