Create OCT29.438PM
DOES COPYING THE ARRAY ELEMENT INTO THE VARIABLE CREATE A SEPARATE MEMORY SPACE...std::string array[]{ "peter", "likes", "frozen", "yogurt" };
for (auto element : array) // element will be a copy of the current array element
{
    std::cout << element << ' ';
}
Yes, copying the array element into the variable creates a separate memory space.

When you copy an array element into a variable, you are creating a new copy of the element in a separate memory location. The original array element and the new copy are completely independent of each other. This means that any changes you make to the new copy will not affect the original array element.

In the code snippet you provided, the element variable will be a copy of the current array element. This means that the element variable will have its own memory location, which is separate from the memory location of the array element. Any changes you make to the element variable will not affect the array element.

Here is an example of how copying an array element into a variable creates a separate memory space:

std::string array[]{ "peter", "likes", "frozen", "yogurt" };

std::string element = array[0]; // copy the first array element into the element variable

element = "ice cream"; // change the value of the element variable

std::cout << array[0] << std::endl; // prints "peter"
std::cout << element << std::endl; // prints "ice cream"
Output:

peter
ice cream
As you can see, changing the value of the element variable does not affect the array element. This is because the element variable is a copy of the array element, and it has its own memory location.

I hope this helps! Let me know if you have any other questions.

profile picture
HOW DOES THE AMPERSAND AFTER THE AUTO TYPE DECLARATION DIFFER FROM COPYING THE ARRAY ELEMENT INTO THE VARIABLE...std::string array[]{ "peter", "likes", "frozen", "yogurt" };
for (auto& element: array) // The ampersand makes element a reference to the actual array element, preventing a copy from being made
{
    std::cout << element << ' ';
}
View other drafts





















The ampersand after the auto type declaration in the code snippet for (auto& element: array) tells the compiler to create a reference to the actual array element. This prevents a copy of the array element from being made.

When you declare a variable with the auto& keyword, you are telling the compiler that the variable should be a reference to another variable. This means that the variable will not have its own memory location. Instead, it will share the memory location of the variable to which it refers.

In the code snippet above, the element variable is a reference to the current array element. This means that the element variable does not have its own memory location. Instead, it shares the memory location of the array element. Any changes you make to the element variable will be directly reflected in the array element.

Here is an example of how using the auto& keyword prevents a copy of the array element from being made:

std::string array[]{ "peter", "likes", "frozen", "yogurt" };

for (auto& element: array) // element is a reference to the actual array element
{
    element = "ice cream";
}

std::cout << array[0] << std::endl; // prints "ice cream"
Output:

ice cream
As you can see, changing the value of the element variable within the for loop also changes the value of the array element. This is because the element variable is a reference to the array element.

I hope this helps! Let me know if you have any other questions.