Create 10.29.658pm
Accumulators
We have already seen a for loop that uses an accumulator variable. This approach can be used with array for aggregate calculations, like to sum or multiply all values in an array.

In this example, we will add together all of the points scored by each player on a sports team. The sum will represent the total points scored by the team.  We will first do this with a range-based for loop (for each loop) and then with a traditional for loop.

Range-Based For Loop:

        int sum = 0; //accumulator variable
	int mylist[] = {8 , 12, 3, 4, 12, 9, 8}};
	for (int player : mylist)
	{
		cout << player << endl;
		sum = sum + player;
	}

	cout << "range based for loop sum: " << sum << endl;
Traditional For Loop

	cout << "Traditional for loop:" << endl;
	sum = 0; //reset sum
        int mylist[] = {8 , 12, 3, 4, 12, 9, 8}};
        for (int i = size - 1; i >=0; i--)
	{
		cout << mylist[i] << endl;
		sum = sum + mylist[i];
	}

	cout << "Traditional for loop sum: " << sum << endl;
Sometimes you need more than one condition to solve a problem: one condition to control the loop, and other conditions to execute or skip instructions at each iteration of the loop. For example, let's say you need to do something to the values of an array, but only if those values meet some condition. In this example, we want to add together all positive numbers in the array, and skip the negative numbers (if any). We will use the range-based for loop for this example and embed an if statement in the body of the loop. The for condition will handle the loop iterations and the if statement will decide which values in the array to sum up.