Create 10.29.728pm
Sometimes you need more than one condition to solve a problem: one condition to control the loop, and other conditions to execute or skip instructions at each iteration of the loop. For example, let's say you need to do something to the values of an array, but only if those values meet some condition. In this example, we want to add together all positive numbers in the array, and skip the negative numbers (if any). We will use the range-based for loop for this example and embed an if statement in the body of the loop. The for condition will handle the loop iterations and the if statement will decide which values in the array to sum up.