Create OCT30.201PM
Iterative Search Algorithm
In this next example, we need to search for a specific name in the array.  If you were to look for a name in a stack of business cards, you might follow these steps for each card to find the card you are looking for.

Here are the steps:

look at the first card
decide if it has the name you are looking for (search key)
display a match if found or skip displaying anything if it is not a match.  (embedded if)
repeat for each value in the array (loop condition)
Because you need to make a decision for each card, you will need a loop to look at each card, so the loop condition needs to begin at the first element of the array, look at each element, then stop at the last element of the array.  Because you need to decide if a specific card has the name you are looking for, you need to embed an if statement in the body of the loop that compares the value of the current element to a search key value that you are looking for.