Create 10.30.352pm
Comparing Two Arrays
This last example compares the contents of two different string arrays. The steps are similar, but with some notable differences.  We will be working with two arrays instead of one and instead of comparing each array value to a known search key value, we need to compare each value in one array to each value in another array.

Here are the steps:

look at the first value of array #1 and the first value of array #2
decide if the first value of array #1 matches the first value of array #2 (embedded if)
repeat for each value in both arrays (loop condition)
Note that this example compares each individual value in both arrays in order. We are not actually comparing the two arrays as a whole.  It is possible that the two arrays have the exact same values, but in a different order.  This code will not be able to determine that.  It only decides if the first value in one array matches the first value in the other array, then compares the second value of each array and so on. 

       string mynames[] = { "Joe", "Sam", "Sue", "Lou" };
	string myothernames[] = { "Joe", "K